Refactor augment reality to navigate route sample
