+++ Shared/Samples/Augment reality to navigate route/AugmentRealityToNavigateRouteView.swift
@@ -14,245 +14,242 @@
 
 import ArcGIS
 import ArcGISToolkit
-import AVFoundation
+import CoreLocation
 import SwiftUI
 
 @MainActor
 struct AugmentRealityToNavigateRouteView: View {
-    /// The data model for the selected route.
-    @StateObject private var routeDataModel = RouteDataModel()
-    /// A Boolean value indicating whether the route planner is showing.
-    @State private var isShowingRoutePlanner = true
-    /// The location datasource that is used to access the device location.
-    @State private var locationDataSource = SystemLocationDataSource()
-    /// A scene with an imagery basemap.
-    @State private var scene = Scene(basemapStyle: .arcGISImagery)
-    /// The elevation surface set to the base surface of the scene.
-    @State private var elevationSurface: Surface = {
-        let elevationSurface = Surface()
-        elevationSurface.navigationConstraint = .unconstrained
-        elevationSurface.opacity = 0
-        elevationSurface.backgroundGrid.isVisible = false
-        return elevationSurface
-    }()
-    /// The elevation source with elevation service URL.
-    @State private var elevationSource = ArcGISTiledElevationSource(url: URL(string: "https://elevation3d.arcgis.com/arcgis/rest/services/WorldElevation3D/Terrain3D/ImageServer")!)
-    /// The graphics overlay containing a graphic.
-    @State private var graphicsOverlay = GraphicsOverlay()
-    /// The status text displayed to the user.
-    @State private var statusText = "Adjust calibration before starting."
-    /// A Boolean value indicating whether the use is navigatig the route.
-    @State private var isNavigating = false
-    /// The result of the route selected in the route planner view.
-    @State private var routeResult: RouteResult?
+    /// The view model for the map view in the sample.
+    @StateObject private var model = MapModel()
     
-    init() {
-        elevationSurface.addElevationSource(elevationSource)
-        scene.baseSurface = elevationSurface
-    }
+    /// The error shown in the error alert.
+    @State private var error: Error?
     
     var body: some View {
-        if isShowingRoutePlanner {
-            RoutePlannerView(isShowing: $isShowingRoutePlanner)
-                .onDidSelectRoute { routeGraphic, routeResult  in
-                    self.routeResult = routeResult
-                    graphicsOverlay = makeRouteOverlay(
-                        routeResult: routeResult,
-                        routeGraphic: routeGraphic
-                    )
-                }
-                .task {
-                    try? await elevationSource.load()
-                }
-        } else {
-            VStack(spacing: 0) {
-                WorldScaleSceneView { _ in
-                    SceneView(scene: scene, graphicsOverlays: [graphicsOverlay])
-                }
-                .calibrationButtonAlignment(.bottomLeading)
-                .onCalibratingChanged { isPresented in
-                    scene.baseSurface.opacity = isPresented ? 0.6 : 0
-                }
-                .task {
-                    try? await locationDataSource.start()
-                    
-                    for await location in locationDataSource.locations {
-                        try? await routeDataModel.routeTracker?.track(location)
-                    }
-                }
-                .overlay(alignment: .top) {
-                    Text(statusText)
-                        .multilineTextAlignment(.center)
-                        .frame(maxWidth: .infinity, alignment: .center)
-                        .padding(8)
-                        .background(.regularMaterial, ignoresSafeAreaEdges: .horizontal)
-                }
-                .onDisappear {
-                    Task { await locationDataSource.stop() }
-                }
-                Divider()
-            }
-            .toolbar {
-                ToolbarItemGroup(placement: .bottomBar) {
-                    Button("Start") {
-                        isNavigating = true
-                        Task {
-                            do {
-                                try await startNavigation()
-                            } catch {
-                                print("Failed to start navigation.")
-                            }
-                        }
-                    }
-                    .disabled(isNavigating)
-                }
-            }
-        }
-    }
-    
-    /// Creates a graphics overlay and adds a graphic (with solid yellow 3D tube symbol)
-    /// to represent the route.
-    private func makeRouteOverlay(routeResult: RouteResult, routeGraphic: Graphic) -> GraphicsOverlay {
-        let graphicsOverlay = GraphicsOverlay()
-        graphicsOverlay.sceneProperties.surfacePlacement = .absolute
-        let strokeSymbolLayer = SolidStrokeSymbolLayer(
-            width: 1.0,
-            color: .yellow,
-            lineStyle3D: .tube
+        MapView(
+            map: model.map,
+            graphicsOverlays: model.graphicsOverlays
         )
-        let polylineSymbol = MultilayerPolylineSymbol(symbolLayers: [strokeSymbolLayer])
-        let polylineRenderer = SimpleRenderer(symbol: polylineSymbol)
-        graphicsOverlay.renderer = polylineRenderer
-        
-        if let originalPolyline = routeResult.routes.first?.geometry {
-            addingElevation(3, to: originalPolyline) { polyline in
-                routeGraphic.geometry = polyline
-                graphicsOverlay.addGraphic(routeGraphic)
-            }
-        }
-        
-        return graphicsOverlay
-    }
-    
-    /// Densify the polyline geometry so the elevation can be adjusted every 0.3 meters,
-    /// and add an elevation to the geometry.
-    ///
-    /// - Parameters:
-    ///   - z: A `Double` value representing z elevation.
-    ///   - polyline: The polyline geometry of the route.
-    ///   - completion: A completion closure to execute after the polyline is generated with success or not.
-    private func addingElevation(
-        _ z: Double,
-        to polyline: Polyline,
-        completion: @escaping (Polyline) -> Void
-    ) {
-        if let densifiedPolyline = GeometryEngine.densify(polyline, maxSegmentLength: 0.3) as? Polyline {
-            let polylineBuilder = PolylineBuilder(spatialReference: densifiedPolyline.spatialReference)
-            Task { @MainActor in
-                for part in densifiedPolyline.parts {
-                    for point in part.points {
-                        async let elevation = try await elevationSurface.elevation(at: point)
-                        let newPoint = await GeometryEngine.makeGeometry(from: point, z: try elevation + z)
-                        // Put the new point 3 meters above the ground elevation.
-                        polylineBuilder.add(newPoint)
-                    }
+        .onSingleTapGesture { _, mapPoint in
+            Task {
+                do {
+                    try await model.addRouteGraphic(for: mapPoint)
+                } catch {
+                    self.error = error
                 }
-                completion(polylineBuilder.toGeometry())
-            }
-        } else {
-            completion(polyline)
-        }
-    }
-    
-    /// Starts navigating the route.
-    private func startNavigation() async throws {
-        guard let routeResult else { return }
-        let routeTracker = RouteTracker(
-            routeResult: routeResult,
-            routeIndex: 0,
-            skipsCoincidentStops: true
-        )
-        guard let routeTracker else { return }
-        
-        routeTracker.voiceGuidanceUnitSystem = Locale.current.measurementSystem == .us
-        ? .imperial
-        : .metric
-        
-        routeDataModel.routeTracker = routeTracker
-        
-        do {
-            try await routeDataModel.routeTask.load()
-        } catch {
-            throw error
-        }
-        
-        if routeDataModel.routeTask.info.supportsRerouting,
-           let reroutingParameters = ReroutingParameters(
-            routeTask: routeDataModel.routeTask,
-            routeParameters: routeDataModel.routeParameters
-           ) {
-            do {
-                try await routeTracker.enableRerouting(using: reroutingParameters)
-            } catch {
-                throw error
             }
         }
-        
-        statusText = "Navigation will start."
-        await startTracking()
-    }
-    
-    /// Starts monitoring multiple asynchronous streams of information.
-    private func startTracking() async {
-        await withTaskGroup(of: Void.self) { group in
-            group.addTask { await trackStatus() }
-            group.addTask { await routeDataModel.trackVoiceGuidance() }
-        }
-    }
-    
-    /// Monitors the asynchronous stream of tracking statuses.
-    ///
-    /// When new statuses are delivered, update the route's traversed and remaining graphics.
-    private func trackStatus() async {
-        guard let routeTracker = routeDataModel.routeTracker else { return }
-        for await status in routeTracker.$trackingStatus {
-            guard let status else { continue }
-            switch status.destinationStatus {
-            case .notReached, .approaching:
-                if let route = routeResult?.routes.first {
-                    let currentManeuver = route.directionManeuvers[status.currentManeuverIndex]
-                    statusText = currentManeuver.text
+        .locationDisplay(model.locationDisplay)
+        .task {
+            Task {
+                do {
+                    try await model.setup()
+                } catch {
+                    self.error = error
                 }
-            case .reached:
-                statusText = "You have arrived!"
-            @unknown default:
-                break
             }
         }
+        .overlay(alignment: .top) {
+            instructionText
+        }
+        .toolbar {
+            ToolbarItemGroup(placement: .bottomBar) {
+                toolbarButtons
+            }
+        }
+        .onAppear {
+            model.statusText = "Tap to place a start point."
+        }
+        .errorAlert(presentingError: $error)
+    }
+    
+    /// The buttons in the bottom toolbar.
+    @ViewBuilder private var toolbarButtons: some View {
+        Spacer()
+        NavigationStack {
+            NavigationLink {
+                ARRouteSceneView(model: model.sceneModel)
+            } label: {
+                Image(systemName: "camera")
+                    .imageScale(.large)
+            }
+            .disabled(!model.didSelectRoute)
+        }
+        
+        Spacer()
+        Button {
+            model.reset()
+            model.statusText = "Tap to place a start point."
+        } label: {
+            Image(systemName: "trash")
+                .imageScale(.large)
+        }
+        .disabled(!model.didSelectRouteStop && !model.didSelectRoute)
+    }
+    
+    /// The instruction text in the overlay.
+    private var instructionText: some View {
+        Text(model.statusText)
+            .multilineTextAlignment(.center)
+            .frame(maxWidth: .infinity, alignment: .center)
+            .padding(8)
+            .background(.thinMaterial, ignoresSafeAreaEdges: .horizontal)
     }
 }
 
-extension AugmentRealityToNavigateRouteView {
+private extension AugmentRealityToNavigateRouteView {
+    // MARK: Map Model
+    
+    /// A view model for this example.
     @MainActor
-    class RouteDataModel: ObservableObject {
-        /// An AVSpeechSynthesizer for text to speech.
-        let speechSynthesizer = AVSpeechSynthesizer()
-        /// The route task that solves the route using the online routing service, using API key authentication.
-        let routeTask = RouteTask(url: URL(string: "https://route-api.arcgis.com/arcgis/rest/services/World/Route/NAServer/Route_World")!)
-        /// The parameters for route task to solve a route.
-        var routeParameters = RouteParameters()
-        /// The route tracker.
-        @Published var routeTracker: RouteTracker?
-        /// The route result.
-        @Published var routeResult: RouteResult?
+    class MapModel: ObservableObject {
+        /// A map with an imagery basemap style.
+        let map = Map(basemapStyle: .arcGISImagery)
         
-        /// Monitors the asynchronous stream of voice guidances.
-        func trackVoiceGuidance() async {
-            guard let routeTracker = routeTracker else { return }
-            for try await voiceGuidance in routeTracker.voiceGuidances {
-                speechSynthesizer.stopSpeaking(at: .word)
-                speechSynthesizer.speak(AVSpeechUtterance(string: voiceGuidance.text))
+        /// The map's location display.
+        let locationDisplay: LocationDisplay = {
+            let locationDisplay = LocationDisplay()
+            locationDisplay.autoPanMode = .recenter
+            return locationDisplay
+        }()
+        
+        /// The graphics overlay for the stops.
+        private let stopGraphicsOverlay = GraphicsOverlay()
+        
+        /// A graphic overlay for route graphics.
+        private let routeGraphicsOverlay: GraphicsOverlay = {
+            let overlay = GraphicsOverlay()
+            overlay.renderer = SimpleRenderer(
+                symbol: SimpleLineSymbol(style: .solid, color: .yellow, width: 5)
+            )
+            return overlay
+        }()
+        
+        /// The map's graphics overlays.
+        var graphicsOverlays: [GraphicsOverlay] {
+            return [stopGraphicsOverlay, routeGraphicsOverlay]
+        }
+        /// A point representing the start of navigation.
+        private var startPoint: Point? {
+            didSet {
+                let stopSymbol = PictureMarkerSymbol(image: .stopA)
+                let startStopGraphic = Graphic(geometry: startPoint, symbol: stopSymbol)
+                stopGraphicsOverlay.addGraphic(startStopGraphic)
             }
+        }
+        /// A point representing the destination of navigation.
+        private var endPoint: Point? {
+            didSet {
+                let stopSymbol = PictureMarkerSymbol(image: .stopB)
+                let endStopGraphic = Graphic(geometry: endPoint, symbol: stopSymbol)
+                stopGraphicsOverlay.addGraphic(endStopGraphic)
+            }
+        }
+        
+        /// A Boolean value indicating whether a route stop is selected.
+        var didSelectRouteStop: Bool {
+            startPoint != nil || endPoint != nil
+        }
+        
+        /// A Boolean value indicating whether a route is selected.
+        var didSelectRoute: Bool {
+            startPoint != nil && endPoint != nil
+        }
+        
+        /// The view model for scene view in the sample.
+        let sceneModel = SceneModel()
+        
+        /// The status text displayed to the user.
+        @Published var statusText = ""
+        
+        deinit {
+            Task {
+                /// Stop the location data source.
+                await locationDisplay.dataSource.stop()
+            }
+        }
+        
+        /// Performs important tasks including setting up the location display, creating route parameters,
+        /// and loading the scene elevation source.
+        func setup() async throws {
+            try await startLocationDisplay()
+            try await makeParameters()
+            try await sceneModel.loadElevationSource()
+        }
+        
+        /// Starts the location display to show user's location on the map.
+        func startLocationDisplay() async throws {
+            // Request location permission if it has not yet been determined.
+            let locationManager = CLLocationManager()
+            if locationManager.authorizationStatus == .notDetermined {
+                locationManager.requestWhenInUseAuthorization()
+            }
+            
+            locationDisplay.dataSource = sceneModel.locationDataSource
+            
+            // Start the location display to zoom to the user's current location.
+            try await locationDisplay.dataSource.start()
+        }
+        
+        /// Creates walking route parameters.
+        func makeParameters() async throws {
+            let parameters = try await sceneModel.routeTask.makeDefaultParameters()
+            
+            if let walkMode = sceneModel.routeTask.info.travelModes.first(where: { $0.name.contains("Walking") }) {
+                parameters.travelMode = walkMode
+                parameters.returnsStops = true
+                parameters.returnsDirections = true
+                parameters.returnsRoutes = true
+                sceneModel.routeParameters = parameters
+            }
+        }
+        
+        /// Adds a route graphic for the selected route using a given start and end point.
+        /// - Parameter mapPoint: The map point for the route start or end point.
+        func addRouteGraphic(for mapPoint: Point) async throws {
+            if startPoint == nil {
+                startPoint = mapPoint
+                statusText = "Tap to place destination."
+            } else if endPoint == nil {
+                endPoint = mapPoint
+                sceneModel.routeParameters.setStops(makeStops())
+                
+                Task {
+                    do {
+                        let routeResult = try await sceneModel.routeTask.solveRoute(
+                            using: sceneModel.routeParameters
+                        )
+                        if let firstRoute = routeResult.routes.first {
+                            let routeGraphic = Graphic(geometry: firstRoute.geometry)
+                            routeGraphicsOverlay.addGraphic(routeGraphic)
+                            sceneModel.routeResult = routeResult
+                            try await sceneModel.makeRouteOverlay()
+                            statusText = "Tap camera to start navigation."
+                        }
+                    } catch {
+                        throw error
+                    }
+                }
+            }
+        }
+        
+        /// Creates the start and destination stops for the navigation.
+        private func makeStops() -> [Stop] {
+            guard let startPoint, let endPoint else { return [] }
+            let stop1 = Stop(point: startPoint)
+            stop1.name = "Start"
+            let stop2 = Stop(point: endPoint)
+            stop2.name = "Destination"
+            return [stop1, stop2]
+        }
+        
+        /// Resets the start and destination stops for the navigation.
+        func reset() {
+            routeGraphicsOverlay.removeAllGraphics()
+            stopGraphicsOverlay.removeAllGraphics()
+            sceneModel.routeParameters.clearStops()
+            startPoint = nil
+            endPoint = nil
         }
     }
 }
